Move ColorProperties and FieldProperties into separate package.
