+++ pixi/src/main/java/org/openpixi/pixi/ui/panel/Particle2DPanel.java
@@ -32,6 +32,8 @@ import javax.swing.JButton;
 import org.openpixi.pixi.physics.Simulation;
 import org.openpixi.pixi.physics.particles.IParticle;
 import org.openpixi.pixi.ui.SimulationAnimation;
+import org.openpixi.pixi.ui.panel.properties.ColorProperties;
+import org.openpixi.pixi.ui.panel.properties.FieldProperties;
 import org.openpixi.pixi.ui.util.FrameRateDetector;
