+++ pixi/src/main/java/org/openpixi/pixi/ui/panel/properties/FieldProperties.java
@@ -1,4 +1,4 @@
-package org.openpixi.pixi.ui.panel;
+package org.openpixi.pixi.ui.panel.properties;
 
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
